Fix URL to code in github.
Apparently, the tensorflow code on github has been reorganized such that older projects like KeypointNet are under the archive branch. As a result, the old link is broken.

This PR points to the code in the archive branch. Perhaps there is a better solution, but for now this should at least fix the broken link.